User pages restriction

+++ base/views.py
@@ -1,6 +1,8 @@
 from urllib import response
+from django.http import HttpResponse
 from django.shortcuts import render, redirect 
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -16,6 +18,9 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -57,6 +62,7 @@ def room (request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url = 'login')
 def createRoom(request):
     form =RoomForm()
     if request.method == 'POST':
@@ -68,9 +74,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = 'login')
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Not allowed !!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -80,8 +90,12 @@ def updateRoom(request, pk):
     context ={'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = 'login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id = pk)
+    if request.user != room.host:
+        return HttpResponse('Not allowed !!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect ('home')
